requirements.txt
pyttsx3==2.98
Pillow==11.0.0
SpeechRecognition==3.11.0
wikipedia==1.4.0
requests==2.32.3
pyjokes==0.8.3
pyaudio==0.2.14